+++ eventcatalog/src/pages/docs/llm/llms-full.txt.ts
@@ -3,8 +3,20 @@ import config from '@config';
 import type { APIRoute } from 'astro';
 import fs from 'fs';
 import { isCustomDocsEnabled } from '@utils/feature';
+import { addSchemaToMarkdown } from '@utils/llms';
 
-type AllowedCollections = 'events' | 'commands' | 'queries' | 'services' | 'domains' | 'teams' | 'users' | 'customPages';
+type AllowedCollections =
+  | 'events'
+  | 'commands'
+  | 'queries'
+  | 'services'
+  | 'domains'
+  | 'teams'
+  | 'users'
+  | 'customPages'
+  | 'channels'
+  | 'entities'
+  | 'flows';
 
 const events = await getCollection('events');
 const commands = await getCollection('commands');
@@ -13,6 +25,9 @@ const services = await getCollection('services');
 const domains = await getCollection('domains');
 const teams = await getCollection('teams');
 const users = await getCollection('users');
+const entities = await getCollection('entities');
+const channels = await getCollection('channels');
+const flows = await getCollection('flows');
 
 const customDocs = await getCollection('customPages');
 
@@ -29,6 +44,9 @@ export const GET: APIRoute = async ({ params, request }) => {
     ...domains,
     ...teams,
     ...users,
+    ...entities,
+    ...channels,
+    ...flows,
   ];
 
   if (isCustomDocsEnabled()) {
@@ -38,7 +56,18 @@ export const GET: APIRoute = async ({ params, request }) => {
   const content = resources
     .map((item) => {
       if (!item.filePath) return '';
-      return fs.readFileSync(item.filePath, 'utf8');
+
+      let file = fs.readFileSync(item.filePath, 'utf8');
+
+      try {
+        // Try and add the schemas to the resource
+        // @ts-ignore
+        file = addSchemaToMarkdown(item, file);
+      } catch (error) {
+        // just skip the resource if it has no schema
+      }
+
+      return file;
     })
     .join('\n');